pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "meshagent-tools"
description = "Tools for Meshagent"
requires-python = ">=3.13"
license = "Apache-2.0"
keywords = []
dependencies = [
  "pyjwt~=2.10",
  "pytest~=8.4",
  "pytest-asyncio~=0.26",
  "meshagent-api~=0.25.0",
  "aiohttp[speedups]~=3.13.0",
  "opentelemetry-distro~=0.54b1"
]
dynamic = ["version", "readme"]

[project.urls]
Documentation = "https://docs.meshagent.com"
Website = "https://www.meshagent.com"
Source = "https://www.meshagent.com"

[tool.setuptools.packages.find]
include = ["meshagent.*"]

[tool.setuptools.package-data]
"meshagent.tools" = ["py.typed", "*.pyi", "**/*.pyi", "**/*.js"]

[tool.setuptools.dynamic]
version = {attr = "meshagent.tools.version.__version__"}
readme = { file = "README.md", content-type = "text/markdown" }
